+++ CHANGELOG.md
@@ -5,15 +5,16 @@
 #### Bugs
 * Fix #6892: rolling().restart() doesn't remove preexistent pod template annotations
 * Fix #6908: The Default annotation and JsonProperty default value should accept JSON values
+* Fix #6906: Knative VolatileTime should be serialized as String
 
 #### Improvements
-
 * Fix #6863: ensuring SerialExecutor does not throw RejectedExecutionException to prevent unnecessary error logs
 * Fix #6763: (crd-generator) YAML output customization
 
 #### Dependency Upgrade
 
 #### New Features
+* Fix #6827: (crd-generator) Add CRDPostProcessor to process generated CRDs before they are written out
 
 #### _**Note**_: Breaking changes